Pull out more tests
