+++ kate/src/gridgen.rs
@@ -376,7 +376,7 @@ fn round_up_to_multiple(input: usize, multiple: NonZeroUsize) -> usize {
 	n_multiples.saturating_mul(multiple.get())
 }
 
-fn pad_to_bls_scalar(a: impl AsRef<[u8]>) -> Result<BlsScalar, Error> {
+pub(crate) fn pad_to_bls_scalar(a: impl AsRef<[u8]>) -> Result<BlsScalar, Error> {
 	if a.as_ref().len() > DATA_CHUNK_SIZE {
 		todo!()
 	}
@@ -496,215 +496,5 @@ mod tests {
 	fn test_get_block_dims(size: usize) -> Dimensions
 where {
 		get_block_dims(size, 4, 256, 256).unwrap()
-	}
-}
-
-#[cfg(test)]
-mod consistency_tests {
-	use super::*;
-	use crate::testnet;
-	use dusk_plonk::prelude::PublicParameters;
-	use hex_literal::hex;
-	use kate_grid::Grid;
-	use kate_recovery::com::reconstruct_extrinsics;
-	use kate_recovery::data::Cell as DCell;
-	use kate_recovery::index::AppDataIndex;
-	use kate_recovery::matrix::Position as DPosition;
-	use proptest::prelude::*;
-	use proptest::{collection, sample::size_range, strategy::Strategy};
-	use rand::distributions::Uniform;
-	use rand::prelude::Distribution;
-
-	pub(crate) fn pp() -> PublicParameters {
-		testnet::public_params(da_types::BlockLengthColumns(256))
-	}
-	#[test]
-	fn newapi_test_flatten_block() {
-		let extrinsics: Vec<AppExtrinsic> = vec![
-			AppExtrinsic {
-				app_id: 0.into(),
-				data: (1..=29).collect(),
-			},
-			AppExtrinsic {
-				app_id: 1.into(),
-				data: (1..=30).collect(),
-			},
-			AppExtrinsic {
-				app_id: 2.into(),
-				data: (1..=31).collect(),
-			},
-			AppExtrinsic {
-				app_id: 3.into(),
-				data: (1..=60).collect(),
-			},
-		];
-
-		let expected_dims = Dimensions::new_unchecked(16, 1);
-		let evals =
-			EvaluationGrid::from_extrinsics(extrinsics, 4, 256, 256, Seed::default()).unwrap();
-
-		let expected_index = [(0.into(), 0), (1.into(), 2), (2.into(), 4), (3.into(), 6)]
-			.into_iter()
-			.map(|(app_id, start)| DataLookupIndexItem { app_id, start })
-			.collect::<Vec<_>>();
-
-		let expected_lookup = DataLookup {
-			size: 9,
-			index: expected_index,
-		};
-
-		assert_eq!(evals.lookup, expected_lookup, "The layouts don't match");
-		assert_eq!(
-			evals.dims, expected_dims,
-			"Dimensions don't match the expected"
-		);
-
-		let expected_data = hex!("04740102030405060708090a0b0c0d0e0f101112131415161718191a1b1c1d00800000000000000000000000000000000000000000000000000000000000000004780102030405060708090a0b0c0d0e0f101112131415161718191a1b1c1d001e80000000000000000000000000000000000000000000000000000000000000047c0102030405060708090a0b0c0d0e0f101112131415161718191a1b1c1d001e1f80000000000000000000000000000000000000000000000000000000000004f00102030405060708090a0b0c0d0e0f101112131415161718191a1b1c1d001e1f202122232425262728292a2b2c2d2e2f303132333435363738393a3b3c00800000000000000000000000000000000000000000000000000000000000000076a04053bda0a88bda5177b86a15c3b29f559873cb481232299cd5743151ac004b2d63ae198e7bb0a9011f28e473c95f4013d7d53ec5fbc3b42df8ed101f6d00e831e52bfb76e51cca8b4e9016838657edfae09cb9a71eb219025c4c87a67c004aaa86f20ac0aa792bc121ee42e2c326127061eda15599cb5db3db870bea5a00ecf353161c3cb528b0c5d98050c4570bfc942d8b19ed7b0cbba5725e03e5f000b7e30db36b6df82ac151f668f5f80a5e2a9cac7c64991dd6a6ce21c060175800edb9260d2a86c836efc05f17e5c59525e404c6a93d051651fe2e4eefae281300");
-
-		let data = evals
-			.evals
-			.inner()
-			.iter()
-			.flat_map(|s| s.to_bytes())
-			.collect::<Vec<_>>();
-		assert_eq!(data, expected_data, "Data doesn't match the expected data");
-	}
-
-	#[test]
-	fn newapi_test_extend_data_matrix() {
-		// This test expects this result in column major
-		let expected_result = vec![
-			hex!("000102030405060708090a0b0c0d0e0f101112131415161718191a1b1c1d1e00"),
-			hex!("bc1c6b8b4b02ca677b825ec9dace9aa706813f3ec47abdf9f03c680f4468555e"),
-			hex!("7c7d7e7f808182838485868788898a8b8c8d8e8f909192939495969798999a00"),
-			hex!("c16115f73784be22106830c9bc6bbb469bf5026ee80325e403efe5ccc3f55016"),
-			hex!("1f202122232425262728292a2b2c2d2e2f303132333435363738393a3b3c3d00"),
-			hex!("db3b8aaa6a21e9869aa17de8f9edb9c625a05e5de399dc18105c872e6387745e"),
-			hex!("9b9c9d9e9fa0a1a2a3a4a5a6a7a8a9aaabacadaeafb0b1b2b3b4b5b6b7b8b900"),
-			hex!("e080341657a3dd412f874fe8db8ada65ba14228d07234403230e05ece2147016"),
-			hex!("3e3f404142434445464748494a4b4c4d4e4f505152535455565758595a5b5c00"),
-			hex!("fa5aa9c9894008a6b9c09c07190dd9e544bf7d7c02b9fb372f7ba64d82a6935e"),
-			hex!("babbbcbdbebfc0c1c2c3c4c5c6c7c8c9cacbcccdcecfd0d1d2d3d4d5d6d7d800"),
-			hex!("ff9f533576c2fc604ea66e07fba9f984d93341ac26426322422d240b02348f16"),
-			hex!("5d5e5f606162636465666768696a6b6c6d6e6f707172737475767778797a7b00"),
-			hex!("197ac8e8a85f27c5d8dfbb26382cf80464de9c9b21d81a574e9ac56ca1c5b25e"),
-			hex!("d9dadbdcdddedfe0e1e2e3e4e5e6e7e8e9eaebecedeeeff0f1f2f3f4f5f6f700"),
-			hex!("1ebf725495e11b806dc58d261ac918a4f85260cb45618241614c432a2153ae16"),
-		]
-		.into_iter()
-		.map(|e| BlsScalar::from_bytes(e.as_slice().try_into().unwrap()).unwrap())
-		.collect::<Vec<_>>()
-		.into_column_major(4, 4)
-		.unwrap()
-		.to_row_major();
-
-		let block_dims = Dimensions::new_unchecked(4, 2);
-		let scalars = (0..=247)
-			.collect::<Vec<u8>>()
-			.chunks_exact(DATA_CHUNK_SIZE)
-			.flat_map(pad_to_bls_scalar)
-			.collect::<Vec<_>>();
-
-		let grid = EvaluationGrid {
-			lookup: DataLookup::default(),
-			evals: scalars
-				.into_row_major(block_dims.width(), block_dims.height())
-				.unwrap(),
-			dims: block_dims,
-		};
-		let extend = grid.extend_columns(2).unwrap();
-
-		assert_eq!(extend.evals.inner(), expected_result.inner());
-	}
-
-	fn app_extrinsic_strategy() -> impl Strategy<Value = AppExtrinsic> {
-		(
-			any::<u32>(),
-			any_with::<Vec<u8>>(size_range(1..2048).lift()),
-		)
-			.prop_map(|(app_id, data)| AppExtrinsic {
-				app_id: app_id.into(),
-				data,
-			})
-	}
-
-	fn app_extrinsics_strategy() -> impl Strategy<Value = Vec<AppExtrinsic>> {
-		collection::vec(app_extrinsic_strategy(), size_range(1..16)).prop_map(|xts| {
-			let mut new_xts = xts;
-			new_xts.sort_by(|a1, a2| a1.app_id.cmp(&a2.app_id));
-			new_xts
-		})
-	}
-
-	fn sample_unique(rng: &mut impl Rng, n_samples: usize, n: usize) -> Vec<usize> {
-		let mut sampled = vec![];
-		let u = Uniform::from(0..n);
-		while sampled.len() < n_samples || sampled.len() < n {
-			let t = u.sample(rng);
-			if !sampled.contains(&t) {
-				sampled.push(t)
-			}
-		}
-		sampled
-	}
-
-	fn app_data_index_from_lookup(lookup: &DataLookup) -> AppDataIndex {
-		AppDataIndex {
-			size: lookup.size,
-			index: lookup.index.iter().map(|e| (e.app_id.0, e.start)).collect(),
-		}
-	}
-
-	proptest! {
-	#![proptest_config(ProptestConfig::with_cases(5))]
-	#[test]
-	fn newapi_test_build_and_reconstruct(exts in app_extrinsics_strategy())  {
-		let grid = EvaluationGrid::from_extrinsics(exts.clone(), 4, 256, 256, Seed::default()).unwrap().extend_columns(2).unwrap();
-		let gref = &grid;
-		let dims = &grid.dims;
-		//let (layout, commitments, dims, matrix) = par_build_commitments(
-		//	BlockLengthRows(64), BlockLengthColumns(16), 32, xts, Seed::default()).unwrap();
-		const RNG_SEED: Seed = [42u8; 32];
-		let mut rng = ChaChaRng::from_seed(RNG_SEED);
-		let cells = (0..dims.width())
-			.flat_map(move |x| {
-				sample_unique(&mut rng, dims.height()/2, dims.height())
-					.into_iter()
-					.map(move |y| {
-						kate_recovery::data::DataCell {
-							position: kate_recovery::matrix::Position { row: y as u32, col: x as u16 },
-							data: gref.evals.get(x, y).unwrap().to_bytes()
-						}
-					}).collect::<Vec<_>>()
-			}).collect::<Vec<_>>();
-		let index = app_data_index_from_lookup(&grid.lookup);
-		let bdims = kate_recovery::matrix::Dimensions::new(dims.height() as u16, dims.width() as u16).unwrap();
-		let reconstructed = reconstruct_extrinsics(&index, &bdims, cells).unwrap();
-		for (result, xt) in reconstructed.iter().zip(exts) {
-			prop_assert_eq!(result.0, *xt.app_id);
-			prop_assert_eq!(result.1[0].as_slice(), &xt.data);
-		}
-
-		let pp = pp();
-		let polys = grid.make_polynomial_grid().unwrap();
-		let commitments = polys.commitments(pp.commit_key()).unwrap();
-		let indices = (0..dims.width()).flat_map(|x| (0..dims.height()).map(move |y| (x, y))).collect::<Vec<_>>();
-
-		// Sample some number 10 of the indices, all is too slow for tests...
-		let mut rng = ChaChaRng::from_seed(RNG_SEED);
-		let sampled = Uniform::from(0..indices.len()).sample_iter(&mut rng).take(10).map(|i| indices[i]);
-		for (x, y) in sampled {
-			let cell = Cell { row: (y as u32).into(), col: (x as u32).into() };
-			let proof = polys.proof(pp.commit_key(), &cell).unwrap();
-			let mut content = [0u8; 80];
-			content[..48].copy_from_slice(&proof.to_bytes()[..]);
-			content[48..].copy_from_slice(&grid.evals.get(x, y).unwrap().to_bytes()[..]);
-
-			let dcell = DCell{position: DPosition { row: y as u32, col: x as u16 }, content };
-			let verification =  kate_recovery::proof::verify(&pp, &bdims, &commitments[y].to_bytes(),  &dcell);
-			prop_assert!(verification.is_ok());
-			prop_assert!(verification.unwrap());
-		}
-	}
 	}
 }
